Move configs to separate module
